+++ weblate/formats/models.py
@@ -146,6 +146,7 @@ class FormatsConf(AppConf):
         "weblate.formats.ttkit.XWikiPagePropertiesFormat",
         "weblate.formats.ttkit.XWikiFullPageFormat",
         "weblate.formats.ttkit.TBXFormat",
+        "weblate.formats.ttkit.FluentFormat",
     )
 
     class Meta: